+++ frontend/src/components/FlowConfig/index.ts
@@ -1,0 +1,3 @@
+export { default as FlowContextMenu } from './FlowContextMenu'
+export { default as TransformEditor } from './TransformEditor'
+export { default as CacheEditor } from './CacheEditor'